chore: split involved parsing logic for reverse lookups into parseWords file
This tidies up the index.ts file and makes it simpler to make changes in the reverse lookups logic.

Also updated the readme file to include the push github action details.
