+++ src/index.ts
@@ -1,10 +1,8 @@
 import fs from 'fs';
 import { cleanCharacters } from './cleanCharacters';
-import { cleanDefinitionTestingFiles, parseDefinition, printParseResult } from './parseDefinition';
+import { cleanDefinitionTestingFiles } from './parseDefinition';
 import { cleanWordTestingFiles, separateWords } from './separateWords';
-import { populateReverseLookupForWordVariant } from './createReverseLookup';
-import { PhraseWordLookups, buildPhraseWordLookups } from './buildPhraseWordLookups';
-import { WordData, WordDataRaw, WordReferenceData, WordVariant, WordVariantRaw } from './wordDataTypes';
+import { parseWords } from './parseWords';
 
 // Define the input and output file paths.
 const dictionaryDataPath = './data/gutenbergWebstersDictionaryCleaned.txt';
@@ -25,57 +23,11 @@ const dictionaryData = cleanCharacters(rawDictionaryData);
 // Separate the words and definition sections so we can parse them further.
 const {wordIdToRawWord, wordIdList} = separateWords(dictionaryData);
 
-// Find the spellings of all the available words, including words with spaces,
-// and populate a lookup based on word id for all available words.
-const phraseWordLookups: PhraseWordLookups = buildPhraseWordLookups(wordIdList, wordIdToRawWord);
+// Parse the words into the format we need to populate the word lookup,
+// including the reverse lookups, into a db.
+const { wordIdToWord, wordIdToWordReferenceData } = parseWords(wordIdToRawWord, wordIdList);
 
-// Process each word into its variant definitions,
-// populating a reverse lookup as we go through each definition.
-const wordIdToWord: {[index: string]: WordData} = {};
-const wordIdToReferenceWordIds: Map<string, string[]> = new Map();
-
-wordIdList.forEach(wordId => {
-  const wordDataRaw: WordDataRaw = wordIdToRawWord[wordId];
-
-  const wordData: WordData = {
-    id: wordDataRaw.id,
-    spellingsString: wordDataRaw.spellingsString,
-    spellings: wordDataRaw.spellings,
-    variants: [],
-  };
-
-  // console.log('parsing word: ', word)
-  wordDataRaw.variants.forEach((variant: WordVariantRaw) => {
-    const wordVariant = parseDefinition(wordData.spellingsString, variant);
-    wordData.variants.push(wordVariant);
-    populateReverseLookupForWordVariant(wordData.id, wordData.spellings, wordVariant.rawData, phraseWordLookups, wordIdToReferenceWordIds);
-  });
-
-  wordIdToWord[wordData.id] = wordData;
-});
-
-// Populate reverse lookup reference data.
-const wordIdToWordReferenceData: {[index: string]: WordReferenceData} = {};
-
-// Find words that reference this word.
-wordIdList.forEach(wordId => {
-  const wordReferences = wordIdToReferenceWordIds.get(wordId);
-  if (wordReferences) {
-    if (wordReferences && wordReferences.length > 5000) {
-      // These have so many references that they aren't useful,
-      // so don't store them.
-      // console.log(' ', wordIdToWord[wordId]?.spellingsString)
-    } else {
-      wordIdToWordReferenceData[wordId] = {
-        wordId: wordId,
-        referenceWordIds: wordReferences
-      }
-    }
-  }
-});
-
-printParseResult();
-
+// Write out the final word data.
 fs.writeFileSync(wordsDataPath, JSON.stringify(Object.values(wordIdToWord), null, 2));
 fs.writeFileSync(wordsReferencesPath, JSON.stringify(Object.values(wordIdToWordReferenceData), null, 2));
 fs.writeFileSync(wordsListPath, JSON.stringify(wordIdList, null, 2));
